+++ src/main/resources/thingl/shaders/regular/sdf_text.vert
@@ -38,10 +38,10 @@ void main() {
     gl_Position = u_ProjectionMatrix * worldPosition;
 
     GlyphData glyphData = glyphDatas[gl_VertexID / 4];
-    TextData textData = textDatas[glyphData.textureAndTextIndex & 0x7FFFFFF];
+    TextData textData = textDatas[glyphData.textureAndTextIndex & uint(0x7FFFFFF)];
 
     v_TexCoords = i_TexCoords;
-    v_TextureIndex = (glyphData.textureAndTextIndex >> 27) & 31;
+    v_TextureIndex = (glyphData.textureAndTextIndex >> 27) & uint(31);
     v_TextColor = unpackUnorm4x8(textData.textColor);
     v_OutlineColor = unpackUnorm4x8(textData.outlineColor);
     v_Smoothing = textData.smoothing;
